requirements/dev.txt
-r base.txt
django-debug-toolbar==1.2.1
# Required by django-debug-toolbar
sqlparse==0.1.11

coverage==3.7.1
pylint==1.1.0
# Required by pylint
logilab-common==0.61.0
astroid==1.0.1

PyYAML==3.10
# Fabric==1.8.2
# Required by Fabric
# paramiko==1.13.0
# pycrypto==2.6.1
# ecdsa==0.11

flake8==2.2.5
  pyflakes==0.8.1
  pep8==1.5.7
  mccabe==0.3

factory_boy==2.4.1

selenium==2.45.0

transifex-client==0.11

requests-mock==0.7.0

# to build Sphinx project docs
Sphinx==1.3.5
  Jinja2==2.8
  Pygments==2.1.1
  docutils==0.12
  MarkupSafe==0.23
  sphinx-rtd-theme==0.1.9
